refactor: remove sql do liquibase para que seja gerado automaticamente com jpa

+++ backend/src/main/java/br/com/digital/flavor/backend/order_item/OrderItem.java
@@ -18,7 +18,7 @@ public class OrderItem {
     private OrderItemId id;
 
     @Column(nullable = false)
-    private int quantity = 1;
+    private Short quantity = 1;
 
     @ManyToOne
     @MapsId("orderId")
